refactor(cfp): address PR review comments
- remove findings.md from version control
- replace manual _get_event_or_404 with Django's get_object_or_404
- extract _is_event_organizer into IsEventOrganizer permission class
- rename app from cfp to cfps to follow plural naming convention
- inherit CFPSubmissionSerializer from WritableNestedModelSerializer
- remove redundant create method; simplify update to delegate to super
