+++ cfps/apps.py
@@ -7,4 +7,4 @@ class CfpConfig(AppConfig):
     """App config for the CFP app."""
 
     default_auto_field = "django.db.models.BigAutoField"
-    name = "cfp"
+    name = "cfps"
