+++ speakwise/settings/base.py
@@ -62,7 +62,7 @@ LOCAL_APPS = [
     "feedbacks",
     "organizations",
     "speakerrequests",
-    "cfp",
+    "cfps",
 ]
 
 INSTALLED_APPS = DJANGO_APPS + THIRD_PARTY_APPS + LOCAL_APPS
