+++ speakwise/urls.py
@@ -29,7 +29,7 @@ urlpatterns = [
     path("api/", include("feedbacks.urls", namespace="feedbacks")),
     path("api/", include("organizations.urls", namespace="organizations")),
     path("api/", include("speakerrequests.urls", namespace="speakerrequests")),
-    path("api/", include("cfp.urls", namespace="cfp")),
+    path("api/", include("cfps.urls", namespace="cfp")),
 ]
 urlpatterns += static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
 urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
